Added support for uploading remote images.
I keep my images in an object store in the cloud, and wanted to display
them in Art Mode on my TV, so this commit adds support for image URL
uploading.

I also encountered some timeouts on the larger files, so I also added
code to read an upload the images in chunks. My non-rigorous testing
found that 64K was a good chunk size.